update readme (#380)
