+++ .env.example
@@ -1,0 +1,12 @@
+NEXT_PUBLIC_ORGANIZATION_NAME=""
+NEXT_PUBLIC_PROJECT_NAME=""
+NEXT_PUBLIC_BACKEND_API_BASE_URL=
+NEXT_PUBLIC_HOTJAR_ID=
+NEXT_PUBLIC_AMPLITUDEANALYTICS_TOKEN=
+NEXT_PUBLIC_TAWK_PROPERTY_ID=
+NEXT_PUBLIC_TAWK_WEIGHT_ID=
+NEXT_PUBLIC_DISCORD_CDN=
+NEXT_PUBLIC_GITBOOK_URL=
+NEXT_PUBLIC_PROJECT_ID=
+NEXT_PUBLIC_APP_DEVELOPER_PUBLIC_ADDRESS=
+NEXT_PUBLIC_IS_MAINNET=
